Added a filterbar + Handle virtually hidden files
This commit consists of two parts that, contrary to appearances, are closely related to each other:

(1) The filterbar is tab-aware but is cleared on changing directory in the same tab. Although it's a fixed part of the GUI, this can be changed later if needed.

(2) Making files "virtually" hidden is a feature that Nemo/Nautilus have and is requested for Dolphin (but ignored by KDE devs) for a long time. The user may want to hide a file without changing or being able to change its name. "lost+found" is a good example.

The commit simply uses the ADT "ProxyFolderModelFilter", which PCMan has included in "proxyfoldermodel.h" for adding extra filters. For virtually hidden files, it reads the file ".hidden" inside the current folder once and hides the listed files from the tabpage but, intentionally, not from the sidepane tree. This exception aside, the virtually hidden files behave like the usual hidden files.