Per-folder saving of sorting and view mode
Separate per-folder savings of sorting and view mode is added to the View sub-menus as check boxes. All settings are saved to and retrieved from `~/.config/pcmanfm-qt/PROFILE/perfolder-settings.conf`. If a box is checked, the current folder will have its own sorting or view mode, which could be changed in the corresponding sub-menu later. If unchecked, the default will be used again.

With this commit, it's possible, for example, to set the wallpaper folder in the thumbnail mode while using the default sorting for it, *and* sort the download folder according to mod. time while it's shown in the default view mode.

The commit uses a nested QHash for saving the folder properties so that they can be read as fast as possible. If a folder is deleted, it will be removed from the list only on quitting.