Make custom pos checkbox work for multiple items
Fixes https://github.com/lxde/pcmanfm-qt/issues/498.

The state of the "Stick to current position" checkbox is corrected for multiple selections. Also when multiplely selected desktop items are right-clicked and the box is checked, all of them stick to their positions; if the box is unchecked, all of them will be sorted according to the sorting order.