Fix showing/hiding hidden files
Fixes https://github.com/lxde/pcmanfm-qt/issues/533.

The bugs that are fixed by this patch:

(1) The status message wasn't changed when hidden files were shown/hidden by context menu (in contrast to `Ctrl+H`).
(2) Showing hidden files wasn't remembered globally.
(3) Sometimes, the setting for showing hidden files wasn't saved for customized folders either.

Current behavior:

(1) If hidden files are shown/hidden inside a folder *without* customization, the setting will be remembered globally for all folders *without* customization.
(2) If hidden files are shown/hidden inside a folder *with* customization, the setting will be remembered only for that folder and not globally.

The behavior implemented by this patch is usually expected for hidden files, as is the case with well-known file managers.

IMHO, this behavior is better to be generalized to other view settings too and it's unfortunate that pcmanfm-qt has no way to inform the user that a folder is customized other than by toggling checkboxes in menus -- but that's another matter to discuss.