Warnings (#625)
* Silent unused parameter warnings

They are so many... that it's easy to miss other "important" warnings.

* Removes unused variable

A artifact I believe.

* Reorder member initialization list

Silents a warning.
From cppreference:

The order of member initializers in the list is irrelevant: the actual
order of initialization is as follows:
1) If the constructor is for the most-derived class, virtual base classes
	are initialized in the order in which they appear in depth-first
	left-to-right traversal of the base class declarations
	(left-to-right refers to the appearance in base-specifier lists)
2) Then, direct base classes are initialized in left-to-right order as
	they appear in this class's base-specifier list
3) Then, non-static data members are initialized in order of declaration
	in the class definition.
4) Finally, the body of the constructor is executed

(Note: if initialization order was controlled by the appearance in the
member initializer lists of different constructors, then the destructor
wouldn't be able to ensure that the order of destruction is the reverse of
the order of construction)

* Construct a MountOperation instance in a proper way

MountOperation(bool interactive = true, QWidget* parent = 0)
We were converting the this pointer to bool.

* Don't compare signed and unsigned integrals

Use size_t type for the loop index as history.size() returns it.
Converting an int (current), to unsigned is valid and well defined. There's
no issue in the comparison, even if current is negative.

* Don't compare signed and unsigned values

size_t is guaranteed to be an unsigned. The comparison is safe because it's
only done if index is positive.