Optionally select newly created files
Closes https://github.com/lxde/pcmanfm-qt/issues/638.

This depends on and follows https://github.com/lxde/libfm-qt/pull/170.

With a new option (unchecked by default) at the end of Preferences → Behavior → File Operations, newly created files are selected and scrolled to appropriately when they are new enough -- whether they are created by pcmanfm-qt or in any other way.

The above-mentioned option is for Desktop too and can be enabled/disabled on-the-fly.

Also, the case of changing Desktop folder is covered.