Always drop into the cell behind cursor
Closes https://github.com/lxqt/pcmanfm-qt/issues/728

Drop desktop items into the cell behind the cursor, regardless of the cursor position inside the cell.

Previously, if the cursor was in the right half of the cell, the item would be dropped into the next cell on the right, and a similar thing happened with the right bottom half.