Added split view
Closes https://github.com/lxqt/pcmanfm-qt/issues/104

The split view can be toggled by `F6`. Switching between view frames can be done by the `Tab` key. Also, when the single tab of a split view frame is closed, the split mode is changed to the simple one.

For now, the split view has only two frames but the code is written in such way that more frames will be possible later.

Please note that the "full-width tab bar" was incompatible with the code of split view and so, is removed. IMHO, preserving it would require an unmaintainable code.