No desktop drop indicator on self
When a folder is dragged over itself on Desktop, there should be no drop indicator (rectangle) because, on Desktop, dropping on self means sticking to the current position. The same is true for the Trash item.

Previously, when a Desktop folder was dragged and dropped on itself, the drop ndicator remained intact, unless another desktop item was dragged.