Removed QDesktopWidget
Closes https://github.com/lxqt/pcmanfm-qt/issues/865

`QDesktopWidget` is obsolete and some of its methods don't work correctly. This patch removes it from the code by using `QScreen` methods directly.

As a result, problems with desktop update are also fixed, especially with virtual screens.

P.S. I made the PR because the patch was ready but it could wait until after 0.14.