Corrected the logic of sorting
Fixes https://github.com/lxqt/pcmanfm-qt/issues/887

When a tab page's sorting is changed in whatever way, remember it.

In this way, there will be no difference between changing the sorting (or showing hidden files) from the context menu and from the menu-bar. Also, because of the sane logic, all hassles related to showing hidden files should be eliminated.