Open selected files by pressing Enter
Closes https://github.com/lxqt/pcmanfm-qt/issues/149

If at least one of the selected items has focus, all of them will be opened by pressing `Enter`. But if their number is more than 20, a confirmation dialog will be first shown (with "No" as its default button).

This feature is especially good -- and maybe needed -- when files have different mime types (but assigned to the same app) because, in that case, they can't be opened by right clicking.