Removed WM menu from desktop
Because:
(1) Only some X11 WMs have desktop context menus, while LXQt is WM-agnostic;
(2) It hid our FM desktop context menu; and
(3) X11 codes should be kept only when really useful because they are an obstacle to preparing for Wayland.

NOTE: "Advanced" is preserved for the tab name because an extra item is added to it outside LXQt.

Closes https://github.com/lxqt/pcmanfm-qt/issues/931