Prevent future crash if a new sort column is added in libfm-qt
If a new sort column is added to detailed list view in libfm-qt but there's no corresponding action in pcmanfm-qt's sort menu, pcmanfm-qt will crash. This patch prevents that.

Such a scenario is possible in 2 ways:

(1) The change may have been made to libfm-qt but not yet to pcmanfm-qt.
(2) Although not likely, we might really not want to add an action corresponding to a new sort column.

Anyways, pcmanfm-qt shouldn't crash.