More stable position for sticky Desktop Trash
Desktop Trash is a special drop target. It will be a little annoying if it is pushed when other items are dropped near it or inside a small empty space of its cell. This patch provides a more stable Trash in two ways:

 1. It prevents a sticky Trash to be moved when other items are dropped near it. Only explicit movements are possible, i.e., by dragging the trash icon.

 2. It trashes dropped items once the drop point is inside the Trash cell — even if it's inside an empty space of that cell.

NOTE: `QListView` ignores empty spaces of cells but our subclassed `FolderViewListView` doesn't and that's a good feature, especially because selection corners are made possible by it.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1007