Try to have a current item inside view as far as possible
Previously, there might be no current item in the view. That might confuse users when they pressed keys like `End`, `PgDn` or `Right` because those keys would select the first item at first and only then would do their expected jobs.

The patch also selects the unselected current item if *the filter-bar is transient and the first pressed key is Space* (previously, the first Space was always put into the transient filter-bar but now, it isn't typed if there is an unselected current item and if typing is done inside the view). This behavior is more consistent with other file managers.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1072