Added an option for single window mode (#1097)
With the option enabled, folders will be opened in new tabs of an existing window (the last focused or created window) if an external app calls pcmanfm-qt, the command-line is used, or folders are clicked on desktop.

The option is disabled by default.

The user could still open folders in new windows by using the context menu or adding the `-n/--new-window` command-line option. Actually, `-n/--new-window` makes sense only after this change.

Also, the pointer to the last active window is changed to a guarded pointer. An ordinary pointer was a potential cause of crash but, luckily, it couldn't cause a crash before this feature.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1094 and fixes https://github.com/lxqt/pcmanfm-qt/issues/1096