Prevented total freeze with auto-mounting of encrypted volumes
Fixes https://github.com/lxqt/pcmanfm-qt/issues/1106

If the user tries to mount a volume in the short time before auto-mounting is started — and that's especially possible with an encrypted volume — a complete freeze might happen because `MountOperation::wait()` uses a local `QEventLoop`. Therefore, we need to start auto-mounting only after all probable GUI mount operations are finished.

The above-mentioned scenario was practically impossible with the desktop mode and so, LXQt users didn't encounter it. But it would be possible if pcmanfm-qt was used outside an LXQt session.