Fixed regression in restoring vertical scroll position (#1178)
Restoring the vertical scroll position after reloading a folder was a nice feature that was damaged after https://github.com/lxqt/pcmanfm-qt/commit/e6065e345825bfa733f816c7e65c5c40a17e7107. The reason was that the `autoScroll` property of the item view is `true` and so, giving the focus to the first item made the view scroll to the top.

This patch fixes the regression by restoring the vertical scroll position only after setting the current item.