Guarantee a correct cursor shape on opening a folder (#1180)
Previously, in the single-click mode and on opening a folder, the cursor kept its shape (e.g., a pointing hand) regardless of whether there was an item under it or not — until it was moved a little.

NOTE: For now, this PR closes https://github.com/lxqt/libfm-qt/issues/561. Later, a similar solution could be added to libfm-qt's `filedialog.cpp` but, IMO, pcmanfm-qt is more important.