Added an option to open desktop folders in default file manager
The option is in the first tab of Desktop Preferences and is unchecked by default. It has effect when desktop folders are launched in the usual way, i.e., by left clicking, and the text label is explicit about that. Middle/right clicking triggers the default behavior.

NOTE: The option is added because the user may want to work with another file manager and use pcmanfm-qt only for drawing desktop; there is no other practical scenario. Therefore, the default should be the old behavior.

Closes https://github.com/lxqt/lxqt/issues/1823