Added an option to open desktop folders in default file manager
The option is in the first tab of Desktop Preferences and is unchecked by default. It has effect when desktop folders are launched in the usual way, i.e., by left clicking, and the text label is explicit about that. Middle/right clicking triggers the default behavior.

NOTE: For user convenience, the home shortcut will also be opened by the default file manager if the option is checked but the other shortcuts are always opened by pcmanfm-qt because other file managers may not be able to handle them.

Closes https://github.com/lxqt/lxqt/issues/1823