Always focus visible view on opening new tab
Previously, when the option "Switch to newly opened tab" was checked, the visible view got focused automatically but, *without that option*, it might remain unfocused (e.g., when the new tab was opened by middle clicking a side-pane bookmark).

This patch guarantees that the visible view is always focused after opening new tabs.