Add initial 'spatial mode' option
Add an option for a 'spatial mode' (might be a more user friendly
description here) in the preferences.

This enables the storing of window position and size in extended
attributes for the particular folder open. The folders are opened
in new windows and will remember the position and size if you move
and resize them.

The 'navigation' elements (side panel, URL bar and the tab bar)
are hidden while in spatial mode. We probably want to add an
'Open Parent' action to the menu as well while in spatial mode (like
Tracker has).

Signed-off-by: Chris Moore <chris@mooreonline.org>