Store view mode and ordering of a folder in the metadata
We now store the folder view mode, sort column, sort order
and folders first option in the metadata for that folder.
We retrieve and apply the metadata on opening the folder
window again.

Filter enabled/disabled is not yet stored due to complications
with the global settings of this option being affected leading
to undesired behaviour at the moment (e.g. if you set the filter
on for one window and off for another, all following windows that
haven't had the metadata saved with continue with the last option
set by the metadata).

I'm not sure the 'case sensitive' option seems to do anything, but
we do store/set this one too...

Signed-off-by: Chris Moore <chris@mooreonline.org>