Add .DirInfo metadata storage for saving to non-extattr locations
We save in the JSON format specification as detailed in this post:
https://github.com/helloSystem/hello/issues/79#issuecomment-795793702

So far we are only saving four 'Window' attributes.

The logic is that the .DirInfo file is loaded and saved for a path when
the metadata is accessed, but the loaded values are only used when the
extended attributes are not present (e.g. when unpacked from an archive).

There is an option in the preferences dialog to enable this feature and
this is enabled by default.

The .DirInfo file is not written if there is a .DisableDirInfo file at
any level in the path hierarchy, e.g. /home/chris/.DisableDirInfo will
prevent any .DirInfo files being written to any directories under
/home/chris and any tree of subdirectories below that path.

Signed-off-by: Chris Moore <chris@mooreonline.org>