desktopwindow: sort devices first
It's not the nicest way, but it works and I couldn't find another way of sorting within the model. Essentially we stop the 'folder first' option for the desktop, as otherwise the 'mountpoint' devices are shown last after all the folders. Then we sort by owner, because devices happen to have a numeric value for the owner and the rest have the user's name. So they appear first...

Signed-off-by: Chris Moore <chris@mooreonline.org>