Don't save/read custom positions of desktop items constantly
Previously, the custom positions of desktop items were saved to disk with any DND, addition or removal of sticky items. They were also read from disk with every update of the desktop layout.

This patch reads the config file at startup. keeps track of custom positions in the memory, and saves them to the config file only when the app is about to exit.

It will be especially useful when an option is added for making all desktop items sticky, although avoiding too much I/O is always good.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1394