Added "Recent Files" to File menu
It covers non-folder, native files that are launched (non-native files could cause trouble). The number of recent files can be set from 0 (the default) to 50.

On exiting the current session, the list of recent files is saved to a separate config file inside pcmanfm-qt's user config directory. But, if the user clears the list, the config file will be emptied immediately, for the sake of privacy.

NOTE1: This needs https://github.com/lxqt/libfm-qt/commit/6b8d5acd76e7c7eb7707257b006995b17231e72f
NOTE2: It is independent of GVFS' `recent:///`, which is useless to us.