Added search history
This patch requires https://github.com/lxqt/libfm-qt/pull/806.

The history items can be selected by Up and Down arrow keys or by opening the combo menu.

For respecting the user's privacy, the maximum search history is zero (= disabled) by default. It can be from 0 to 50, with separate lists for the name and content entries.

Also, a button is added to Preferences dialog for clearing the search history.

The options are in preferences → Advanced → Search.

Closes https://github.com/lxqt/pcmanfm-qt/issues/1580