Fixed a small problem in tab DND
Drag info should have been reset on releasing mouse button.

The issue this patch fixes can be seen by dragging a tab just a little and releasing it, so that it returns to its previous position. After that, all tabs of that window (except for the last remaining one) could be thrown outside it by dragging them from their close buttons.