Allow users to add terminal commands
Now, the terminal command entered by the user will be added to the list of terminals if the Preferences dialog is accepted (by pressing the OK button or pressing Enter inside the combo).

The format is as shown by the GUI, except for the redundant `%s`, which I didn't remove for the sake of translations — it's harmless. In most cases, just the terminal command is enough (without option).

The user can also right click inside the combo and remove a terminal if added previously or open the user-defined list if existing.

This PR is made based on https://github.com/lxqt/libfm-qt/pull/924 and needs it to work.