Separate multi-screen desktops under Wayland
For avoiding the problems of multi-screen desktops, each screen is given its separate desktop with a separate items layout under Wayland.

X11 isn't touched.

NOTE. IMO, this is the best solution for having multi-screen desktops under Wayland. The reason is a Wayland issue that limits an extended desktop to one screen as soon as any window is shown on a layer. A workaround for that issue was possible, but flickers couldn't be avoided.