Prevent going up with auto-repeated Backspace in transient filter-bar (#1999)
If the transient filter-bar is shown and Backspace is kept pressed, no going-up will happen. But if the transient filter-bar is emptied and Backspace is released, its next press will still cause going up (although without auto-repeating).

This is the only possible compromise between keeping Backspace for going-up and preventing an unwanted going-up with the transient filter-bar.

NOTE: Pressing Backspace isn't a good way of clearing the transient filter-bar because the selected file may change. Pressing Escape is always the best and easiest way. I might add that to pcmanfm-qt's Wiki page.