Handle Wayland screens by using their names (#2006)
The main reason behind this change is that screen setting by using numbers may not work with all Wayland compositors.

With this patch, and on Wayland, the screen names are used for saving the arrangements of desktop items.

No change is made to X11 desktops.

NOTE: This is a backward incompatible change, in the sense that the arrangement of Wayland desktop items will be rest at first. But Wayland users could copy the contents of `desktop-items-0.conf` and paste it into `desktop-items-SCREEN_NAME.conf` after stopping the desktop module. That couldn't be done by the code, because it isn't known to which screen the old config file belonged (the number proves nothing).