Show bulk rename dialog again on failure (#2030)
A failure means one of these: (1) nothing is renamed, (2) an empty string is searched, or (3) an invalid regex is searched. Then the dialog will be shown again with its previous state, after the error message is closed.

Also, as a compromise between different opinions, a group-box is added and all group-boxes are made mutually exclusive. This also simplifies the code a little.