Fixed shortcut removal with multiple desktops (#2060)
With multiple desktops, if shortcuts were removed, error dialogs would be shown. The reason was that the shortcuts were removed in the first instance of desktop with a delay, and the code tried to remove them again in the other instances.

The problem didn't surface until each Wayland screen was given a separate desktop — on X11, there's only one extended desktop.

Fixes https://github.com/lxqt/pcmanfm-qt/issues/2059