Consistent visibility of multi-screen desktop items on Wayland (#2112)
Hiding desktop items is made dependent on Wayland screen. The change is backward-compatible, in the sense that it doesn't affect X11, where all screens share a single extended desktop.

Closes https://github.com/lxqt/pcmanfm-qt/issues/2109

NOTE: The code is written in a way to make it easier to add more backward-compatible, screen-dependent key-value pairs, although that's not needed for now.